fix: hide warning in the logResponse

+++ src/Services/HttpRequestService.php
@@ -184,7 +184,7 @@ class HttpRequestService
      * https://laravel.com/docs/10.x/telescope and
      * https://packagist.org/packages/muhammadhuzaifa/telescope-guzzle-watcher instead
      */
-    protected function logResponse(?int $time = null): void
+    protected function logResponse(?float $time = null): void
     {
         $endTime = (empty($time)) ? null : microtime(true) - $time;
